removed unnecessary files form codegen, changed real_T datatype to float, added float suffix to constants

+++ vitis/software/Baremetal/src/uz/uz_ResonantController/Resonant_Controller_ert_rtw/rtwtypes.h
@@ -65,7 +65,7 @@ typedef double real64_T;
  * Generic type definitions: boolean_T, char_T, byte_T, int_T, uint_T,       *
  *                           real_T, time_T, ulong_T, ulonglong_T.           *
  *===========================================================================*/
-typedef double real_T;    /**< real datatype*/
+typedef float real_T;    /**< real datatype*/
 typedef double time_T;
 typedef unsigned char boolean_T;
 typedef int int_T;
